requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-Cors==4.0.0
python-dotenv==1.0.1
python-telegram-bot==13.15
